Revert "feat(docs): unify page-actions into a single "Copy page" dropdown"
This reverts commit 7770d4542d302a86068c995ca19e57a743fb3b98.
