+++ src/app/docs/[...slug]/page.tsx
@@ -4,6 +4,7 @@ import { MDXRemote } from "next-mdx-remote/rsc";
 import { getAllDocSlugs, getDocBySlug } from "@/lib/mdx";
 import { MDX_COMPONENTS } from "@/components/docs/mdx-components";
 import { EditOnGitHub } from "@/components/docs/EditOnGitHub";
+import { DocPageActions } from "@/components/docs/DocPageActions";
 import { PageActionsMenu } from "@/components/docs/PageActionsMenu";
 
 import remarkGfm from "remark-gfm";
@@ -59,7 +60,8 @@ export default async function DocPage({ params }: PageProps) {
           </div>
 
           <div className="flex flex-wrap items-center gap-3">
-            <PageActionsMenu
+            <PageActionsMenu slug={doc.slug} title={doc.frontmatter.title} markdownContent={doc.content} />
+            <DocPageActions
               slug={doc.slug}
               title={doc.frontmatter.title}
               description={doc.frontmatter.description}
