+++ src/components/docs/PageActionsMenu.tsx
@@ -1,73 +1,21 @@
 "use client";
 
-import { Fragment, useEffect, useRef, useState } from "react";
-import type { LucideIcon } from "lucide-react";
-import {
-  Bot,
-  Check,
-  ChevronDown,
-  Copy,
-  ExternalLink,
-  FileCode2,
-  FileJson,
-  FileText,
-  Github,
-  Loader2,
-  MessageSquare,
-} from "lucide-react";
+import { useEffect, useRef, useState } from "react";
+import { Bot, Check, ChevronDown, Copy, ExternalLink, MessageSquare } from "lucide-react";
 import { logger } from "@/utils/logger";
 import { SITE_URL_FALLBACK } from "@/constants/site";
-import { DOCS_EDIT_BASE } from "@/constants/github";
-import { exportDocMarkdown } from "@/lib/docs/export-doc-markdown";
-import { exportDocJson } from "@/lib/docs/export-doc-json";
-import { exportDocPdf } from "@/lib/docs/export-doc-pdf";
 
 const SITE_URL = process.env.NEXT_PUBLIC_SITE_URL || SITE_URL_FALLBACK;
-
-const MENU_ITEM_CLASS =
-  "w-full flex items-center gap-2 px-4 py-2 text-sm text-content-primary " +
-  "hover:text-theme-primary hover:bg-theme-primary/5 text-left " +
-  "focus-visible:outline-2 focus-visible:outline-theme-primary focus-visible:outline-offset-[-2px]";
 
 interface PageActionsMenuProps {
   slug: string;
   title: string;
-  description?: string;
   markdownContent: string;
 }
 
-/**
- * Groups menu items for divider placement. "connect" is reserved for
- * follow-up issues (Connect with MCP / VSCode / Claude Code / Codex) —
- * appending items with that group inserts a divider before them
- * automatically, with no changes needed elsewhere in this component.
- */
-type MenuItemGroup = "copy" | "export" | "connect";
-
-interface MenuItemBase {
-  id: string;
-  label: string;
-  icon: LucideIcon;
-  group: MenuItemGroup;
-}
-
-interface ButtonMenuItem extends MenuItemBase {
-  kind: "button";
-  onSelect: () => void;
-  busy?: boolean;
-}
-
-interface LinkMenuItem extends MenuItemBase {
-  kind: "link";
-  href: string;
-}
-
-type DocPageMenuItem = ButtonMenuItem | LinkMenuItem;
-
-export function PageActionsMenu({ slug, title, description, markdownContent }: PageActionsMenuProps) {
+export function PageActionsMenu({ slug, title, markdownContent }: PageActionsMenuProps) {
   const [isOpen, setIsOpen] = useState(false);
   const [copied, setCopied] = useState(false);
-  const [isExportingPdf, setIsExportingPdf] = useState(false);
   const containerRef = useRef<HTMLDivElement>(null);
   const triggerRef = useRef<HTMLButtonElement>(null);
   const itemRefs = useRef<Array<HTMLButtonElement | HTMLAnchorElement | null>>([]);
@@ -77,7 +25,6 @@ export function PageActionsMenu({ slug, title, description, markdownContent }: P
   const aiPrompt = `Read ${title} at ${absoluteRawMarkdownUrl} and help me understand it.`;
   const chatGptUrl = `https://chatgpt.com/?q=${encodeURIComponent(aiPrompt)}`;
   const claudeUrl = `https://claude.ai/new?q=${encodeURIComponent(aiPrompt)}`;
-  const githubEditUrl = `${DOCS_EDIT_BASE}/content/docs/${slug}.mdx`;
 
   useEffect(() => {
     if (!isOpen) return;
@@ -137,100 +84,8 @@ export function PageActionsMenu({ slug, title, description, markdownContent }: P
     closeMenu();
   }
 
-  function handleExportMarkdown() {
-    exportDocMarkdown(slug, markdownContent);
-    closeMenu();
-  }
-
-  function handleExportJson() {
-    exportDocJson(slug, title);
-    closeMenu();
-  }
-
-  async function handleExportPdf() {
-    if (isExportingPdf) return;
-
-    setIsExportingPdf(true);
-    try {
-      await exportDocPdf({ slug, title, description });
-    } catch (error) {
-      logger.error("PDF export failed", error);
-    } finally {
-      setIsExportingPdf(false);
-      closeMenu();
-    }
-  }
-
-  const items: DocPageMenuItem[] = [
-    {
-      id: "copy-page",
-      kind: "button",
-      group: "copy",
-      icon: copied ? Check : Copy,
-      label: copied ? "Copied!" : "Copy page",
-      onSelect: handleCopyPage,
-    },
-    {
-      id: "view-markdown",
-      kind: "link",
-      group: "copy",
-      icon: ExternalLink,
-      label: "View as Markdown",
-      href: rawMarkdownUrl,
-    },
-    {
-      id: "open-chatgpt",
-      kind: "link",
-      group: "copy",
-      icon: MessageSquare,
-      label: "Open in ChatGPT",
-      href: chatGptUrl,
-    },
-    {
-      id: "open-claude",
-      kind: "link",
-      group: "copy",
-      icon: Bot,
-      label: "Open in Claude",
-      href: claudeUrl,
-    },
-    {
-      id: "export-markdown",
-      kind: "button",
-      group: "export",
-      icon: FileCode2,
-      label: "Export Markdown",
-      onSelect: handleExportMarkdown,
-    },
-    {
-      id: "export-json",
-      kind: "button",
-      group: "export",
-      icon: FileJson,
-      label: "Export JSON",
-      onSelect: handleExportJson,
-    },
-    {
-      id: "export-pdf",
-      kind: "button",
-      group: "export",
-      icon: isExportingPdf ? Loader2 : FileText,
-      label: isExportingPdf ? "Exporting PDF…" : "Export PDF",
-      onSelect: handleExportPdf,
-      busy: isExportingPdf,
-    },
-    {
-      id: "edit-github",
-      kind: "link",
-      group: "export",
-      icon: Github,
-      label: "Edit on GitHub",
-      href: githubEditUrl,
-    },
-  ];
-
   return (
-    <div className="relative" ref={containerRef} data-pdf-exclude="true">
+    <div className="relative" ref={containerRef}>
       <button
         ref={triggerRef}
         type="button"
@@ -238,7 +93,7 @@ export function PageActionsMenu({ slug, title, description, markdownContent }: P
         aria-haspopup="menu"
         aria-expanded={isOpen}
         aria-controls="page-actions-menu"
-        className="neu-circle h-10 flex items-center gap-2 px-4 text-sm font-medium text-content-secondary hover:text-theme-primary focus-visible:outline-2 focus-visible:outline-theme-primary focus-visible:outline-offset-2 focus-visible:ring-2 focus-visible:ring-theme-primary focus-visible:ring-offset-2 focus-visible:ring-offset-bg-sunken"
+        className="neu-circle h-10 flex items-center gap-2 px-4 text-sm font-medium text-content-secondary hover:text-[#149A9B] focus-visible:outline-2 focus-visible:outline-theme-primary focus-visible:outline-offset-2 focus-visible:ring-2 focus-visible:ring-theme-primary focus-visible:ring-offset-2 focus-visible:ring-offset-bg-sunken"
       >
         {copied ? <Check size={16} /> : <Copy size={16} />}
         <span>{copied ? "Copied!" : "Copy page"}</span>
@@ -251,48 +106,65 @@ export function PageActionsMenu({ slug, title, description, markdownContent }: P
           role="menu"
           aria-label="Page actions"
           onKeyDown={handleMenuKeyDown}
-          className="absolute right-0 top-full mt-2 w-64 rounded-xl z-[100] bg-bg-elevated border border-theme-border/40 shadow-2xl shadow-black/10 py-2 animate-dropdownIn"
+          className="absolute right-0 top-full mt-2 w-56 rounded-xl z-[100] bg-bg-elevated border border-theme-border/40 shadow-2xl shadow-black/10 py-2"
         >
-          {items.map((item, index) => {
-            const Icon = item.icon;
-            const showsDivider = index > 0 && items[index - 1].group !== item.group;
+          <button
+            ref={(el) => {
+              itemRefs.current[0] = el;
+            }}
+            type="button"
+            role="menuitem"
+            onClick={handleCopyPage}
+            className="w-full flex items-center gap-2 px-4 py-2 text-sm text-content-primary hover:text-[#149A9B] hover:bg-theme-primary/5 text-left focus-visible:outline-2 focus-visible:outline-theme-primary focus-visible:outline-offset-[-2px]"
+          >
+            {copied ? <Check size={16} /> : <Copy size={16} />}
+            <span>{copied ? "Copied!" : "Copy page"}</span>
+          </button>
 
-            return (
-              <Fragment key={item.id}>
-                {showsDivider && <div role="separator" className="my-1 h-px bg-theme-border/40" />}
-                {item.kind === "button" ? (
-                  <button
-                    ref={(el) => {
-                      itemRefs.current[index] = el;
-                    }}
-                    type="button"
-                    role="menuitem"
-                    aria-busy={item.busy || undefined}
-                    onClick={item.onSelect}
-                    className={MENU_ITEM_CLASS}
-                  >
-                    <Icon size={16} className={item.busy ? "animate-spin" : undefined} />
-                    <span>{item.label}</span>
-                  </button>
-                ) : (
-                  <a
-                    ref={(el) => {
-                      itemRefs.current[index] = el;
-                    }}
-                    href={item.href}
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    role="menuitem"
-                    onClick={handleExternalLinkClick}
-                    className={MENU_ITEM_CLASS}
-                  >
-                    <Icon size={16} />
-                    <span>{item.label}</span>
-                  </a>
-                )}
-              </Fragment>
-            );
-          })}
+          <a
+            ref={(el) => {
+              itemRefs.current[1] = el;
+            }}
+            href={rawMarkdownUrl}
+            target="_blank"
+            rel="noopener noreferrer"
+            role="menuitem"
+            onClick={handleExternalLinkClick}
+            className="w-full flex items-center gap-2 px-4 py-2 text-sm text-content-primary hover:text-[#149A9B] hover:bg-theme-primary/5 focus-visible:outline-2 focus-visible:outline-theme-primary focus-visible:outline-offset-[-2px]"
+          >
+            <ExternalLink size={16} />
+            <span>View as Markdown</span>
+          </a>
+
+          <a
+            ref={(el) => {
+              itemRefs.current[2] = el;
+            }}
+            href={chatGptUrl}
+            target="_blank"
+            rel="noopener noreferrer"
+            role="menuitem"
+            onClick={handleExternalLinkClick}
+            className="w-full flex items-center gap-2 px-4 py-2 text-sm text-content-primary hover:text-[#149A9B] hover:bg-theme-primary/5 focus-visible:outline-2 focus-visible:outline-theme-primary focus-visible:outline-offset-[-2px]"
+          >
+            <MessageSquare size={16} />
+            <span>Open in ChatGPT</span>
+          </a>
+
+          <a
+            ref={(el) => {
+              itemRefs.current[3] = el;
+            }}
+            href={claudeUrl}
+            target="_blank"
+            rel="noopener noreferrer"
+            role="menuitem"
+            onClick={handleExternalLinkClick}
+            className="w-full flex items-center gap-2 px-4 py-2 text-sm text-content-primary hover:text-[#149A9B] hover:bg-theme-primary/5 focus-visible:outline-2 focus-visible:outline-theme-primary focus-visible:outline-offset-[-2px]"
+          >
+            <Bot size={16} />
+            <span>Open in Claude</span>
+          </a>
         </div>
       )}
     </div>
