+++ src/components/docs/__tests__/PageActionsMenu.test.tsx
@@ -2,12 +2,8 @@ import { describe, it, expect, vi, beforeEach } from "vitest";
 import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { PageActionsMenu } from "../PageActionsMenu";
-import { DOCS_EDIT_BASE } from "@/constants/github";
 
 const loggerError = vi.fn();
-const exportDocMarkdown = vi.fn();
-const exportDocJson = vi.fn();
-const exportDocPdf = vi.fn();
 
 vi.mock("@/utils/logger", () => ({
   logger: {
@@ -15,26 +11,9 @@ vi.mock("@/utils/logger", () => ({
   },
 }));
 
-vi.mock("@/lib/docs/export-doc-markdown", () => ({
-  exportDocMarkdown: (...args: unknown[]) => exportDocMarkdown(...args),
-}));
-
-vi.mock("@/lib/docs/export-doc-json", () => ({
-  exportDocJson: (...args: unknown[]) => exportDocJson(...args),
-}));
-
-vi.mock("@/lib/docs/export-doc-pdf", () => ({
-  exportDocPdf: (...args: unknown[]) => exportDocPdf(...args),
-}));
-
 function renderMenu(markdownContent = "# Title\n\nBody copy.") {
   return render(
-    <PageActionsMenu
-      slug="getting-started"
-      title="Getting Started"
-      description="A getting started guide."
-      markdownContent={markdownContent}
-    />,
+    <PageActionsMenu slug="getting-started" title="Getting Started" markdownContent={markdownContent} />,
   );
 }
 
@@ -52,9 +31,6 @@ async function openMenu(user: ReturnType<typeof userEvent.setup>) {
 describe("PageActionsMenu", () => {
   beforeEach(() => {
     loggerError.mockClear();
-    exportDocMarkdown.mockClear();
-    exportDocJson.mockClear();
-    exportDocPdf.mockReset();
   });
 
   it("does not render the menu until opened", () => {
@@ -62,42 +38,16 @@ describe("PageActionsMenu", () => {
     expect(screen.queryByRole("menu")).not.toBeInTheDocument();
   });
 
-  it("opens the menu on trigger click with the expected menu items, in order", async () => {
+  it("opens the menu on trigger click with the expected menu items", async () => {
     const user = userEvent.setup({ delay: null });
     renderMenu();
 
     await openMenu(user);
 
-    const menuItems = screen.getAllByRole("menuitem");
-    expect(menuItems.map((item) => item.textContent)).toEqual([
-      "Copy page",
-      "View as Markdown",
-      "Open in ChatGPT",
-      "Open in Claude",
-      "Export Markdown",
-      "Export JSON",
-      "Export PDF",
-      "Edit on GitHub",
-    ]);
-  });
-
-  it("renders exactly one divider between the copy/AI group and the export group", async () => {
-    const user = userEvent.setup({ delay: null });
-    renderMenu();
-
-    await openMenu(user);
-
-    const menu = screen.getByRole("menu");
-    const separators = screen.getAllByRole("separator");
-    expect(separators).toHaveLength(1);
-
-    const children = Array.from(menu.children);
-    const separatorIndex = children.indexOf(separators[0]);
-    const exportMarkdownIndex = children.findIndex((child) => child.textContent === "Export Markdown");
-    const openClaudeIndex = children.findIndex((child) => child.textContent === "Open in Claude");
-
-    expect(separatorIndex).toBeGreaterThan(openClaudeIndex);
-    expect(separatorIndex).toBeLessThan(exportMarkdownIndex);
+    expect(screen.getByRole("menuitem", { name: /copy page/i })).toBeInTheDocument();
+    expect(screen.getByRole("menuitem", { name: /view as markdown/i })).toBeInTheDocument();
+    expect(screen.getByRole("menuitem", { name: /open in chatgpt/i })).toBeInTheDocument();
+    expect(screen.getByRole("menuitem", { name: /open in claude/i })).toBeInTheDocument();
   });
 
   it("closes the menu on Escape and returns focus to the trigger", async () => {
@@ -119,7 +69,7 @@ describe("PageActionsMenu", () => {
     const user = userEvent.setup({ delay: null });
     render(
       <div>
-        <PageActionsMenu slug="getting-started" title="Getting Started" markdownContent="content" />
+        <PageActionsMenu slug="getting-started" markdownContent="content" />
         <button type="button">outside</button>
       </div>,
     );
@@ -132,38 +82,42 @@ describe("PageActionsMenu", () => {
     });
   });
 
-  it("moves focus between all 8 items with ArrowDown/ArrowUp, wrapping at both ends", async () => {
+  it("moves focus between items with ArrowDown/ArrowUp", async () => {
     const user = userEvent.setup({ delay: null });
     renderMenu();
 
     await openMenu(user);
 
-    const orderedLabels = [
-      /^copy page$/i,
-      /view as markdown/i,
-      /open in chatgpt/i,
-      /open in claude/i,
-      /export markdown/i,
-      /export json/i,
-      /^export pdf$/i,
-      /edit on github/i,
-    ];
-    const orderedItems = orderedLabels.map((name) => screen.getByRole("menuitem", { name }));
+    const copyItem = screen.getByRole("menuitem", { name: /copy page/i });
+    const viewItem = screen.getByRole("menuitem", { name: /view as markdown/i });
+    const chatGptItem = screen.getByRole("menuitem", { name: /open in chatgpt/i });
+    const claudeItem = screen.getByRole("menuitem", { name: /open in claude/i });
 
-    expect(orderedItems[0]).toHaveFocus();
+    expect(copyItem).toHaveFocus();
 
-    for (let i = 1; i < orderedItems.length; i++) {
-      await user.keyboard("{ArrowDown}");
-      expect(orderedItems[i]).toHaveFocus();
-    }
-
-    // wraps back to the first item
     await user.keyboard("{ArrowDown}");
-    expect(orderedItems[0]).toHaveFocus();
+    expect(viewItem).toHaveFocus();
 
-    // wraps backward to the last item
+    await user.keyboard("{ArrowDown}");
+    expect(chatGptItem).toHaveFocus();
+
+    await user.keyboard("{ArrowDown}");
+    expect(claudeItem).toHaveFocus();
+
+    await user.keyboard("{ArrowDown}");
+    expect(copyItem).toHaveFocus();
+
     await user.keyboard("{ArrowUp}");
-    expect(orderedItems[orderedItems.length - 1]).toHaveFocus();
+    expect(claudeItem).toHaveFocus();
+
+    await user.keyboard("{ArrowUp}");
+    expect(chatGptItem).toHaveFocus();
+
+    await user.keyboard("{ArrowUp}");
+    expect(viewItem).toHaveFocus();
+
+    await user.keyboard("{ArrowUp}");
+    expect(copyItem).toHaveFocus();
   });
 
   it("copies the markdown content to the clipboard and shows a confirmation", async () => {
@@ -177,7 +131,7 @@ describe("PageActionsMenu", () => {
     renderMenu("# Hello\n\nWorld.");
 
     await openMenu(user);
-    await user.click(screen.getByRole("menuitem", { name: /^copy page$/i }));
+    await user.click(screen.getByRole("menuitem", { name: /copy page/i }));
 
     expect(writeText).toHaveBeenCalledWith("# Hello\n\nWorld.");
     await waitFor(() => {
@@ -196,7 +150,7 @@ describe("PageActionsMenu", () => {
     renderMenu();
 
     await openMenu(user);
-    await user.click(screen.getByRole("menuitem", { name: /^copy page$/i }));
+    await user.click(screen.getByRole("menuitem", { name: /copy page/i }));
 
     await waitFor(() => {
       expect(loggerError).toHaveBeenCalled();
@@ -261,133 +215,6 @@ describe("PageActionsMenu", () => {
 
     await waitFor(() => {
       expect(screen.queryByRole("menu")).not.toBeInTheDocument();
-    });
-  });
-
-  it("exports Markdown via exportDocMarkdown and closes the menu", async () => {
-    const user = userEvent.setup({ delay: null });
-    renderMenu("# Hello\n\nWorld.");
-
-    await openMenu(user);
-    await user.click(screen.getByRole("menuitem", { name: /export markdown/i }));
-
-    expect(exportDocMarkdown).toHaveBeenCalledWith("getting-started", "# Hello\n\nWorld.");
-    await waitFor(() => {
-      expect(screen.queryByRole("menu")).not.toBeInTheDocument();
-    });
-  });
-
-  it("exports JSON via exportDocJson and closes the menu", async () => {
-    const user = userEvent.setup({ delay: null });
-    renderMenu();
-
-    await openMenu(user);
-    await user.click(screen.getByRole("menuitem", { name: /export json/i }));
-
-    expect(exportDocJson).toHaveBeenCalledWith("getting-started", "Getting Started");
-    await waitFor(() => {
-      expect(screen.queryByRole("menu")).not.toBeInTheDocument();
-    });
-  });
-
-  it("links Edit on GitHub to the edit URL for the current slug's source file", async () => {
-    const user = userEvent.setup({ delay: null });
-    renderMenu();
-
-    await openMenu(user);
-
-    const link = screen.getByRole("menuitem", { name: /edit on github/i });
-    expect(link).toHaveAttribute("href", `${DOCS_EDIT_BASE}/content/docs/getting-started.mdx`);
-    expect(link).toHaveAttribute("target", "_blank");
-    expect(link).toHaveAttribute("rel", "noopener noreferrer");
-  });
-
-  it("closes the menu when clicking Edit on GitHub", async () => {
-    const user = userEvent.setup({ delay: null });
-    renderMenu();
-
-    await openMenu(user);
-    await user.click(screen.getByRole("menuitem", { name: /edit on github/i }));
-
-    await waitFor(() => {
-      expect(screen.queryByRole("menu")).not.toBeInTheDocument();
-    });
-  });
-
-  describe("Export PDF", () => {
-    it("shows a busy state while pending, keeps the menu open, and calls exportDocPdf with the doc's details", async () => {
-      const user = userEvent.setup({ delay: null });
-      let resolveExport: () => void = () => {};
-      exportDocPdf.mockImplementation(
-        () =>
-          new Promise<void>((resolve) => {
-            resolveExport = resolve;
-          }),
-      );
-
-      renderMenu();
-
-      await openMenu(user);
-      await user.click(screen.getByRole("menuitem", { name: /export pdf/i }));
-
-      expect(exportDocPdf).toHaveBeenCalledWith({
-        slug: "getting-started",
-        title: "Getting Started",
-        description: "A getting started guide.",
-      });
-
-      const busyItem = await screen.findByRole("menuitem", { name: /exporting pdf/i });
-      expect(busyItem).toHaveAttribute("aria-busy", "true");
-      // the menu must stay open while the export is in flight
-      expect(screen.getByRole("menu")).toBeInTheDocument();
-
-      resolveExport();
-
-      await waitFor(() => {
-        expect(screen.queryByRole("menu")).not.toBeInTheDocument();
-      });
-    });
-
-    it("logs an error and clears the busy state when the export fails", async () => {
-      const user = userEvent.setup({ delay: null });
-      exportDocPdf.mockRejectedValue(new Error("export failed"));
-
-      renderMenu();
-
-      await openMenu(user);
-      await user.click(screen.getByRole("menuitem", { name: /export pdf/i }));
-
-      await waitFor(() => {
-        expect(loggerError).toHaveBeenCalledWith("PDF export failed", expect.any(Error));
-      });
-      await waitFor(() => {
-        expect(screen.queryByRole("menu")).not.toBeInTheDocument();
-      });
-    });
-
-    it("ignores repeat clicks while an export is already in flight", async () => {
-      const user = userEvent.setup({ delay: null });
-      let resolveExport: () => void = () => {};
-      exportDocPdf.mockImplementation(
-        () =>
-          new Promise<void>((resolve) => {
-            resolveExport = resolve;
-          }),
-      );
-
-      renderMenu();
-
-      await openMenu(user);
-      const pdfItem = screen.getByRole("menuitem", { name: /export pdf/i });
-      await user.click(pdfItem);
-      await user.click(screen.getByRole("menuitem", { name: /exporting pdf/i }));
-
-      expect(exportDocPdf).toHaveBeenCalledTimes(1);
-
-      resolveExport();
-      await waitFor(() => {
-        expect(screen.queryByRole("menu")).not.toBeInTheDocument();
-      });
     });
   });
 });
